+++ app/src/main/java/com/kamikadze328/mtstetaproject/adapter/movieshort/MovieShortAdapter.kt
@@ -5,7 +5,7 @@ import androidx.recyclerview.widget.ListAdapter
 import com.kamikadze328.mtstetaproject.adapter.movie.MovieCallback
 import com.kamikadze328.mtstetaproject.data.dto.Movie
 
-class MovieShortAdapter(private val click: (id: Int) -> Unit) :
+class MovieShortAdapter(private val click: (id: Long) -> Unit) :
     ListAdapter<Movie, MovieShortViewHolder>(MovieCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
