+++ app/src/main/java/com/kamikadze328/mtstetaproject/presentation/main/CallbackGenreClicked.kt
@@ -1,5 +1,5 @@
 package com.kamikadze328.mtstetaproject.presentation.main
 
 fun interface CallbackGenreClicked {
-    fun onGenreClicked(genreId: Int)
+    fun onGenreClicked(genreId: Long)
 }
